EC-98 Added functions for loading category-buttons on productlist page
